Update invitation.py
Updating email invitation text, and adding variable {org_name} so that it's easy to have the invitation reflect the relevant organisation's name.
This branch stills needs to have the variable implemented elsewhere in the code. But I don't know how to do that – so please feel free to give me a hand with this. :-)